chatbot/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.40
chromadb>=0.4.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
flashrank>=0.2.0
tavily-python>=0.3.0
pdfplumber>=0.10.0
langdetect>=1.0.9
deep-translator>=1.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
